subo todo

+++ src/controllers/makeController.js
@@ -1,4 +1,4 @@
-import { cargarProducto, deleteProductsPorId, modificarProductoPorId } from "../services/funciones.js";
+import { cargarProducto, deleteProductsPorId, modificarProductoPorId, searchProductsPorId2 } from "../services/funciones.js";
 import { HttpResponse } from '../middlewares/errors-handle.js';
 
 const makeController = {};
@@ -19,20 +19,38 @@ makeController.crearProducto = async (req, res, next) => {
 
     const { titulo, descripcion, code, precio, cantidad, marca, categoria, demografia } = req.body;
 
-    const producto = {
-      titulo: titulo,
-      descripcion: descripcion,
-      code: code,
-      precio: precio,
-      cantidad: cantidad,
-      marca: marca,
-      categoria: categoria,
-      demografia: demografia
+    if (req.user.rol === "premium") {
+      const producto = {
+        titulo: titulo,
+        descripcion: descripcion,
+        code: code,
+        precio: precio,
+        cantidad: cantidad,
+        marca: marca,
+        categoria: categoria,
+        demografia: demografia,
+        owner: req.user.email
+      }
+
+      console.log(producto);
+
+      await cargarProducto(producto);
+    } else {
+      const producto = {
+        titulo: titulo,
+        descripcion: descripcion,
+        code: code,
+        precio: precio,
+        cantidad: cantidad,
+        marca: marca,
+        categoria: categoria,
+        demografia: demografia,
+      }
+
+      console.log(producto);
+
+      await cargarProducto(producto);
     }
-
-    console.log(producto);
-
-    await cargarProducto(producto);
 
     res.render("make", { title: "Make handelbars" });
   } catch (error) {
@@ -49,9 +67,31 @@ makeController.borrarProducto = async (req, res, next) => {
 
     console.log(idProducto)
 
-    await deleteProductsPorId(idProducto)
+    if (req.user.rol === "premium") {
 
-    res.render("make", { title: "Make handelbars" });
+      console.log("si");
+
+      const pro = await searchProductsPorId2(idProducto)
+
+      if (req.user.email == pro.owner) {
+
+        await deleteProductsPorId(idProducto)
+
+        res.render("make", { title: "Make handelbars" });
+
+      }
+      else {
+        console.log("no tienes permitido modificar este producto")
+      }
+    } else {
+
+      await deleteProductsPorId(idProducto)
+
+      res.render("make", { title: "Make handelbars" });
+    }
+
+
+
   } catch (error) {
     console.error("Error al crear el usuario:", error.message);
     res.redirect("/log");
@@ -62,11 +102,30 @@ makeController.modificarProducto = async (req, res, next) => {
   try {
     console.log("modificarProducto");
 
-    const { idProducto , campo, valor } = req.body;
+    const { idProducto, campo, valor } = req.body;
 
-    console.log(idProducto)
+    if (req.user.rol === "premium") {
 
-    await modificarProductoPorId(idProducto, campo, valor)
+      const pro = await searchProductsPorId2(idProducto)
+
+      if (req.user.email == pro.owner) {
+
+        console.log(idProducto)
+    
+        await modificarProductoPorId(idProducto, campo, valor)
+
+      }
+      else {
+        console.log("no tienes permitido modificar este producto")
+      }
+
+    } else {
+
+      console.log(idProducto)
+  
+      await modificarProductoPorId(idProducto, campo, valor)
+      }
+
     res.render("make", { title: "Make handelbars" });
   } catch (error) {
     console.error("Error al crear el usuario:", error.message);
